kubam: modified test cases to use names instead of Ids

+++ kubam/app/server/servers.py
@@ -77,6 +77,8 @@ class Servers(object):
         print req
         if not isinstance(req, dict):
             return {"error": "invalid parameters: {0}".format(req)}, Const.HTTP_BAD_REQUEST
+        if not "name" in req:
+            return {"error": "please give the server group name you wish to delete"}, Const.HTTP_BAD_REQUEST
         name = req['name']
         db = YamlDB()
         err, msg = db.delete_server_group(Const.KUBAM_CFG, name)
